Fixed leaks
