+++ Makefile
@@ -24,7 +24,7 @@ PARSING_PATH = mandatory/parsing/
 HEREDOC_PATH= mandatory/heredoc/
 SIGNALS_PATH = mandatory/signals/
 
-SRC = main.c create_env.c delete_env.c
+SRC = main.c create_env.c delete_env.c prompt.c
 EXECUTION = execution.c init.c open.c pid_manager.c fd_manager.c builtins.c
 UTIL = commands.c sort_list.c env_replacer.c transform_env.c utils.c parsing_utils.c \
 	   split_skip_quotes_utils.c split_utils.c dollars_utils.c cpy_txt_utils.c quotes_parser.c ft_realloc.c
